Support "unoffical Bats strict mode"
Part of #20. Addresses sstephenson/bats#171.

Bats is now capable of executing test cases that conform to the
"unoffical Bats strict mode" described by:

- http://redsymbol.net/articles/unofficial-bash-strict-mode/

While the test fixture is straightforward, it took trial and error to
realize a few things. First, `set -u` changed in Bash 4.1-alpha. Chapter
and verse from https://tiswww.case.edu/php/chet/bash/CHANGES:

  This document details the changes between this version,
  bash-4.1-alpha, and the previous version, bash-4.0-release.

  n.  Fixed the behavior of `set -u' to conform to the latest Posix
      interpretation: every expansion of an unset variable except $@ and
      $* will cause the shell to exit.

This almost certainly accounts for the fact that referencing
`"${FOO[@]}"` when `FOO=()` is defined is OK under the latest Bash
4.4.12(1)-release, but breaks under the Bash 3.2.57(1)-release that
ships with macOS. The fix was to do an array length check before each
such operation.

It probably also accounts for the fact that under Bash 3.2.57, no error
output is generated at all for certain unset variable accesses. The
workaround there is to run the test using the latest version of Bash.

The biggest surprise was realizing that no error output is visible when
the unset variable access happens inside a DEBUG trap handler. All
versions of Bash seem susceptible to this.

As a result of these debugging subtleties, the test case contains a
thorough error message to suggest how to proceed when the fixture fails.