Don't truncate skip message when ')' is included.
Addresses sstephenson/bats#240.  When outputting a 'pretty' formatted
output stream, the 'skip' message was being processed with a character
class of [^)].  I couldn't find a reason for this to be the case, so
I've swapped it out to check instead of printable characters, spaces
included ([:print:]).  Please advise if there is a problem with this
approach.

Tested using Bash 4.3.11(1)-release.