Revert "Test correct permission setup by install.sh"
This reverts commit ab52948b74c81a8d3aa4b49d378655da7865a2c0.

The provided test failed on AppVeyor and Travis.
This is very likely a problem of the test implementation itself.

AppVeyor runs on a windows system. So the output of `stat -c '%A'` will very likely be different than for a linux system.

AppVeyor error message:

```
not ok 44 installer only fixes file permissions of installed files
```

Travis fails on alpine linux while calling the install script:

```
not ok 44 installer only fixes file permissions of installed files
```

The problems need to be researched and fixed before adding this test again.