Add test for installer permissions
The test makes sure that install.sh doesn't change the permissions of files other than the installed ones.

After applying this commit, make sure to rebuild your docker image:

```
docker build --tag bats:latest .
```

The change to the docker file was necessary for the Travis CI test to succeed. It also performs a test using the docker image and it wasn't able to find install.sh in that case. However, because we're running tests in different environment, it was inevitable to refer to install.sh using a relative path.

Testing with the git bash in AppVeyor was another challenge. The execution flags of installed and previously existing files needed to be checked. For crossplatform scenarios `find` is the easiest solution for checking the execution permissions.

However, the git bash –  on a Windows system, as performed by the AppVeyor test – would use Windows `find`, which expects a different parameter syntax compared to gnu find. This can be solved by using `/bin/find` in the git bash. See https://stackoverflow.com/questions/21438556/finding-files-in-a-git-bash-terminal#answer-40633578