Show errors when Bash EXIT trap fires without ERR
Closes #72 and #81. Both of those issues deal with the `bats_error_trap`
executing with an incorrect value for `$?`. Bats exited with an error,
but provided no stack trace information to pinpoint its source.

* #72 deals with the inconsistent failure-mode behavior of the `source`
  builtin between Bash versions. I believe it's feasible to update
  `load` with robust error checking around `source` and strongly
  recommend its use in Bats test files instead of using `source`
  directly. I've already opened #80 to track this.

* #81 deals with inconsistent failure-mode behavior of `set -u`,
  particularly when code within a Bats test file itself references an
  unbound variable.

With credit to sstephenson, Bats was smart enough to know and to report
that an error occurred, even when it couldn't tell exactly where the
error came from.

Building on that existing mechanism, this change produces output for
these failure cases even when `bats_error_trap` isn't called with the
correct value for `$?`. It passes all the existing tests under Bash
3.2.57(1)-release and 4.4.19(1)-release. I also timed the original
suites with and without the change to ensure the runtime cost was
negligible.

A couple more notes:

* This change obsoletes the suggestion in #81 that a test case to
  validate testing code running under `set -u` is necessary. `run()`
  disables `set -e` for the code under test to begin with, sidestepping
  the problem with `set -eu` interactions in Bash versions prior to
  4.1.0.

* I've noticed that this mechanism does _not_ work when the problematic
  behavior is in `teardown()`. The change preserves existing behavior,
  the remedy for that issue requires further thinking (and a new issue).