bats_error_trap: Check BATS_TEST_COMPLETED, not $?
This is actually a more reliable way to check for an error than checking
the value of `$?`, since `$?` can sometimes be zero even after an error
occurred.

The updated test cases failed before the code change was applied, and
pass after.

Also updated the comment to try to be clearer about the fact that the
test will generate no failure output when an EXIT trigger happens while
`bats_teardown_trap` executes as the EXIT trap.