Make `printf -v` target variables explicit
Per @sublimino's suggestion in #91, I've made the target of some `printf
-v` statements more explicit by assigning local variable names to
explicit parameters.

I originally did this for statements in `libexec/bats-exec-test` as
well, but since those functions are called so frequently, there was a
slight-yet-noticeable performance impact. Also, many of those functions
are one- or two-liners, so I didn't consider the change to be worth it.