Fix BATS_CWD test case build breakage in Docker
As it turns out, the test case shouldn't assume that `PWD` is a parent
directory of `BATS_TEST_DIRNAME`. Our Docker build, which doesn't set
`/opt/bats` as the `WORKDIR`, broke as a result of that assumption:

  https://travis-ci.org/bats-core/bats-core/jobs/390264918

Now the test case uses a path constructed with `PWD` directly, rather
than one constructed with `BATS_TEST_DIRNAME`.