Ensure last line of output prints without newline
While running my mbland/go-script-bash tests with v1.0.0 (which I
should've tried _before_ v1.0.0), I noticed some failed due to the last
line of output missing. This happened because the `while` loop used to
replace `sed` in #88 would break the loop when the last line didn't end
with a newline.

As it turns out, the fix was pretty easy, thanks to the hints from:

- https://stackoverflow.com/a/14547230
- https://unix.stackexchange.com/a/418067

However, there was a slight performance hit for the existing tests
when the `[[ -n "$line" ]]` conditional appeared as part of the `while`
expression. Instead, this change emits the last line in an `if`
statement after the `while` loop.

I got my mbland/go-script-bash tests to pass by updating the code to
always emit a newline, which was the right thing to do regardless.
However, it seemed still worth handling this condition in Bats itself,
to maintain behavioral parity with v0.4.0.