Add support for recursion for test suites
Closes #84.

Test suites can now place tests in subdirectories.  These are usable
with the '-r' or '--recursive' flag, and recursion will be used when
a directory is given to bats as an argument for testing.  There is no
effect on execution when a test file is passed with this flag included.

Tested on Bash 4.3.11(1)-release and 4.3.33(1)-release