BATS_ROOT: Use `set -P`, remove `PWD` resolution
Today I learned about the `set -P` option, which causes `cd` to set
`PWD` to the physical directory, containing no symlinks. This eliminated
the need for the inner `while` loop to resolve `PWD` symlinks.

The Bash man page wording is a little confusing, but there's a great
example demonstrating the option at:

  https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html

The result is the same, but subshells are only spawned to resolve
target_name links, making the code a little smaller and improving the
performance ever-so-slightly.