Cover more Bash versions with Docker
Revised the Dockerfile to use the official Bash images for improved
version-specific test coverage.  These images are based on Alpine.  Given that
we're only trying to support 3.2.57 and up (because OSX), coverage has been
set to only go that far back (though there are older versions available in
Docker image form if we would like to use them).