Add duration tracking option for each test
In some validation cases, it makes sense to track the duration of each
test to understand which ones are executing quickly and which ones are
executing slowly.  Trends may be built up over time to see how duration
changes given the other changes being tested.

Rebasing from harschware/bats/yaml_durations.

Because this adds a YAML block to the TAP output, the TAP producer needs
to write "TAP version 13" as the first line.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>