libexec/bats: Consolidate, update usage
Per @sublimino's suggestion in #126, `usage()` and `help()` have been
combined into `usage()`. Also updates the format of the `Usage:` stanza
to make it more readable.

Also adds the `BATS_VERSION` environment variable, which seems like a
reasonable item to export. This implementation is in conflict with
the change @jasonkarns made in #124, though as small change to that
PR should bring it in line.