trace: Add tests, optimize, fix error case
This final version will filter out any stack trace values and trace
output where:

- the `DEBUG` trap fires for the function definition itself, resulting
  in redundant --trace output lines (comparing `BASH_COMMAND` against
  `${FUNCNAME[1]}`)

- the `DEBUG` trap fires before the `ERR` trap on error; apparently when
  both functions fire, `$?` refers to the value returned by the last
  command before either trap is fired (checking `$status` before
  emitting the trace output)

The use of `case` patterns instead of multiple tests in an `if`
condition is a bit more compact, and very slightly faster on average.
Hence, I took the opportunity to apply a `case` statement in
`bats_capture_stack_trace` as well.