*: refactor to always use bats-exec-suite
This allows us to remove most of the complicated handling between
bats-exec-test and bats-exec-suite (which had a complicated interplay
due to the fact they both supported multi-test execution).

Instead, multi-test execution is entirely left to bats-exec-suite and
bats-exec-test only handles a single test. This entirely removes the
need for the magic output handling in bats-exec-test -- and opens the
door to suite-wide parallelism in one fell swoop.

The only downside is that bats-preprocess is run multiple times on the
same source file, but a follow-up patch will improve the caching of
pre-processed files such that they will only be processed once.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>