test: correct bats-format-tap-stream test
It appears as though this test didn't actually test what it was supposed
to in 25505bd14324 ("Skip pretty formatting if the first line isn't a
TAP plan"). It was supposed to test that bats-format-tap-stream would
not try to format non-TAP data -- but instead it just showed that the
preamble of test files was executed before printing the TAP header
(something that is not going to be the case with the suite refactor).

Fixes: 25505bd14324 ("Skip pretty formatting if the first line isn't a TAP plan")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>