bats: add support for parallel job execution with GNU parallel
Most well-behaved bats tests (where they aren't writing to shared data
between test executions) are able to be executed in parallel, but to
date bats has always executed them in serial. This is incredibly
wasteful for trivially parallelisable tests, and can result in fairly
inflated test times for no good reason. Parallel exeuction is done
suite-wide, so it makes no difference how the tests are structured.

Unfortunately, because some suites might be implicitly serial (like
bats' own test suite) we cannot make the default mode execute in
parallel. In addition, there is currently no way of tagging that a
*specific* test has to be run in serial (so that the rest can be run in
parallel) -- but that can be done in future work.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>