Update Docker infrastructure for bats
Bats is no longer installed in `/usr/sbin`, bats isn't a sbin candidate. It is
moved to `/usr/local/bin`. This should not have any side effects as
`/usr/local/bin` should be in ones PATH.

Have a default docker-compose.yml for users that they can use to easily play
with docker-compose.

Signed-off-by: Wesley Schwengle <wesley@opndev.io>