Fix the implementation of the run helper function to properly parse lines of output.
The existing behavior is broken, as it relies on the shell to word-split unquoted
variables. This breaks down in several ways, notably that empty lines are silently
dropped, but also because wordsplitting also triggers globbing, so any line that
happens to match a shell glob will be expanded in the resulting array.