Include symlinks during --recursive operation
Previously, to implement recursive operation, bats used find without the -L switch. As a consequence, when the test cases were structured in subdirectories, where some of the folders were symlinks or the testcase itself was a symlink to a testcase, it would be ignored.

The -L option of find follows symlinks and checks the test predicates on the target of the link.