install.sh: consider multilib
It may install files to /usr/lib64 when multilib is enabled. Add an
optional parameter for install.sh to support it.

And add test case for it. It just copied the first test case in
install.bats and pass the second parameter to script install.sh. Then
check files under libdir for both test cases. Add function teardown() in
install.bats as well.

Signed-off-by: Kai Kang <kai.kang@windriver.com>