Fix incorrect line information on test failures
Bats tries hard to tell users the failing line when a test fails. For
example:

     ✗ run tests which consume stdin (see #197)
       (in test file test/bats.bats, line 516)
         `[ "$status" -eq 1 ]' failed

This line reporting is broken for a number of cases. For example, Bats
reports the wrong line in the following cases:

* With older versions of bash (such as the version deployed on macOS
  10.12.6), when a `[[` or `((` check fails.

* When a command succeeds but its success was turned into a failure
  using a `!` prefix.

* When a command inside a function fails, but the failure is handled by
  the caller using `||`, but the failure handler also fails.

See bats_debug_trap's documentation for an explanation of why these
cases behave oddly.

Redesign Bats' interpretation of traps to fix these cases. The new
design is very fragile, but works for more cases than the old design.