Stricter `parallel` Exists Check
There are 2 common implementations of `parallel`: GNU parallel
(https://www.gnu.org/software/parallel/) and moreutils' parallel
(https://manpages.debian.org/testing/moreutils/parallel.1.en.html,
https://linux.die.net/man/1/parallel).

The former is much more feature rich and is the actual intended
dependency of parallel tests in bats.

The latter is far more common on many operating systems (including
Ubuntu).

Currently, when the `parallel.bats` file runs, it checks to see if
`parallel` is available and will skip the test if not. Unfortunately,
this only checks that there is _a_ `parallel` command available -- not
necessarily the correct one. If moreutils parallel is installed, the
skip will not occur and the test will later fail in
"libexec/bats-core/bats-exec-suite", when trying to run `parallel
--keep-order --jobs ...` -- these CLI args are unknown to moreutils
`parallel`.

This PR improves the check such that not _any_ `parallel` command will
work. This check is extended by checking that the `parallel` command
supports `--version`, which distinguishes between GNU parallel
(supported) and moreutils parallel (unsupported).