test/parallel: add a test for negative jobs number arguments
Before commit 48b0d2650a96, -j2 and -j -2 both triggered an infinite
loop. That commit fixed the infinite loop but -j2 and -j -2 are still
parsed the same. In the corresponding code review #657 Martin expressed
an intent to change the parser and parse these differently in the
future.

In commit 0ec3619426f5 Martin added a "no infinite loop" test for
"-j2". However this test will stop testing negative arguments when the
parser changes (the test will also have to be adjusted when the parser
changes).

Add one additional test with a "-j -3" argument that will test a
negative argument forever.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>