Add a --negative-filter option (#1114)
See https://pagure.io/fedora-qa/os-autoinst-distri-fedora/pull-request/389 .
The FAQ claimed "So you could try to invert the regex", but as
far as I can tell, there's no practical way to do this with bash
regexes, if you want to express "run all tests except foo_in_a"
or something like that. You can do it in some regex flavors using
negative lookahead matches, but bash's regex flavor does not seem
to support that, and you can't just stick a `!` on the front of
the regex or anything like that.

So, this adds --negative-filter which just does the opposite of
--filter. The implementation is intentionally 'dumb' because
handling 'what if one or both is not defined?' gets a bit awkward
if you try to handle both in one function.

Signed-off-by: Adam Williamson <awilliam@redhat.com>