bats: stream bats-exec-suite output even with --jobs
The --keep-order argument was included in the call to GNU parallel to the
initial parallelism implementation back in commit da65d1bcecfe ("bats: add
support for parallel job execution with GNU parallel").

My concern at the time was that some parsers of TAP output might
struggle to handle out-of-order results. However, the TAP specification
explicitly states that this is permitted behaviour:

 > Test Points may be output in any order, but any Test Point ID
 > provided must be within the range described by the Plan.

The current state of bats is such that "--jobs" effectively makes bats
test runs completely silent until the very end, which is not very
user-friendly. To resolve this issue, this patch changes the behaviour
to simply stream the output (and adds a --parallel-ordered-output flag
for users that desire the old behaviour).

Note that this does not affect the ordering of per-file tests, as those
are run by a different parallelisation system that would be a bit harder
to add support for out-of-order output. This is left for future work.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>